Remove podIds and clarity refactor (extract method)
I believe by making MuPod Equatable (same pattern as with MuHub) there is no need to track pod ids in a parallel collection in MuDock. I think this could be simplified even further if using a Set as the collection instead of an Array, assuming Swift has a Set implementation, but noot going that far for now.

Also extracted the sub pod building into its own method while I was trying to work out what it did, seemed ok to leave that change in.