MuIdentity tests
I was curious id MuIdentity was thread safe, so I wrote a test to see if it would fail.

It didn't fail, but I'm still not sure if that is because my test isn't aggressive enough or if it is actually thread safe.

I wanted to spawn 100 concurrently executable threads and throw that at it, but the only way I could find to do this with async/await is using the `async let` assignment which requires individually declaring a var for each parallel task and then `await` all of them. I suspect that this is just an insurmountable limitation of the async/await paradigm and that to do a more aggressive test I would need to drop back to NSOperation or GCD or whatever.